refactor: inner class 분리
